refactor: split server.js into app.js & server.js
